+++ src/apisix_client/plugin/models/authentication/__init__.py
@@ -1,0 +1,2 @@
+from apisix_client.plugin.models.authentication.base_auth import BaseAuth, RouteBaseAuth
+from apisix_client.plugin.models.authentication.key_auth import KeyAuth, KeyAuthSettings