+++ src/apisix_client/plugin/models/observability/__init__.py
@@ -1,0 +1,2 @@
+from apisix_client.plugin.models.observability.loggers.clickhouse import ClickhouseLogger
+from apisix_client.plugin.models.observability.loggers.file import FileLogger