+++ src/apisix_client/plugin/models/security/__init__.py
@@ -1,0 +1,1 @@
+from apisix_client.plugin.models.security.consumer_restriction import ConsumerRestriction